Split the finance-data context object out of the provider file
finance-data-context.tsx exported both the FinanceDataContext object and the
FinanceDataProvider component, which tripped react-refresh/only-export-components
(the repo's only lint warning) and would break Fast Refresh for the provider.

Move the context object and its FinanceDataContextValue type into a new
component-free module, finance-data-context-value.ts; the provider file now
exports only the component, and useFinanceData imports the context from the new
module. App and test-utils still import FinanceDataProvider from the same path,
so they're untouched. eslint is now clean (0 warnings); build, format, tests, and
the 100% src/helpers coverage gate stay green.
